requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
docopt==0.6.2
idna==3.6
itsdangerous==2.1.2
jinja2==3.1.3
markupsafe==2.1.5
mysql-connector-python==8.3.0
python-dotenv==1.0.1
requests==2.31.0
sqlalchemy==2.0.25
typing-extensions==4.9.0
urllib3==2.2.0
yarg==0.1.9
pydantic==2.7.0
werkzeug==3.0.2
gunicorn==22.0.0
flask-cors==4.0.0
psycopg2==2.9.9